requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
pandas
google-cloud-storage